+++ spec/auction_fun_core/workers/operations/auction_context/post_auction/winner_operation_job_spec.rb
@@ -2,19 +2,20 @@
 
 require "spec_helper"
 
-RSpec.describe AuctionFunCore::Workers::Operations::AuctionContext::Processor::FinishOperationJob, type: :worker do
+RSpec.describe AuctionFunCore::Workers::Operations::AuctionContext::PostAuction::WinnerOperationJob, type: :worker do
   let(:auction_repository) { AuctionFunCore::Repos::AuctionContext::AuctionRepository.new }
-  let(:auction) { Factory[:auction, :default_running_standard] }
+  let(:auction) { Factory[:auction, :default_standard, :with_winner] }
 
   describe "#perform" do
     subject(:worker) { described_class.new }
 
     context "when params are valid" do
-      it "expect execute auction start operation" do
-        expect { worker.perform(auction.id) }
-          .to change { auction_repository.by_id(auction.id).status }
-          .from("running")
-          .to("finished")
+      it "expect execute winner mailer service" do
+        allow(AuctionFunCore::Workers::Services::Mail::AuctionContext::PostAuction::WinnerMailerJob).to receive(:perform_async)
+
+        worker.perform(auction.id, auction.winner_id)
+
+        expect(AuctionFunCore::Workers::Services::Mail::AuctionContext::PostAuction::WinnerMailerJob).to have_received(:perform_async).once
       end
     end
 
@@ -25,7 +26,7 @@ RSpec.describe AuctionFunCore::Workers::Operations::AuctionContext::Processor::F
       end
 
       it "expect rescue/capture exception and reschedule job" do
-        expect { worker.perform(nil) }.to change(described_class.jobs, :size).from(0).to(1)
+        expect { worker.perform(nil, nil) }.to change(described_class.jobs, :size).from(0).to(1)
 
         expect(AuctionFunCore::Application[:logger]).to have_received(:error).at_least(:once)
       end
@@ -38,7 +39,7 @@ RSpec.describe AuctionFunCore::Workers::Operations::AuctionContext::Processor::F
       end
 
       it "expect raise exception and stop retry" do
-        expect { worker.perform(nil) }.to raise_error(ROM::TupleCountMismatchError)
+        expect { worker.perform(nil, nil) }.to raise_error(ROM::TupleCountMismatchError)
 
         expect(AuctionFunCore::Application[:logger]).to have_received(:error).at_least(:once)
       end